+++ app/src/main/java/com/oreo/ui/custom/sleep/internal/GraphData.kt
@@ -3,7 +3,7 @@ package com.oreo.ui.custom.sleep.internal
 import java.time.LocalDate
 
 data class GraphDataModel(
-    val value1: Int?=null,
-    val value2: Int?=null,
+    val value1: Float?=null,
+    val value2: Float?=null,
     val date: LocalDate,
 )
